Add CWL tests under ARM (#4038)
* Add CWL tests under ARM

* Add abstract class for cluster creation tests

* Add CWL on ARM tests to CI

* Inherit abstract cluster creation class in CWL tests on ARM

* Add needs_env_var decorator back

* Remove unnecessary methods from AbstractAWSAutoscaleTest

* Switch leaderNodeType to t4g.medium

* Remove spaces around TOIL_AWS_KEYNAME

* Install private key for CWL ARM tests

* Add clusterType arg to launchCluster

* Switch _require_env to  needs_env_var

* Create toil in tmp directory

* Increase leaderNodeType size
